changed CTA file name
