--- src/components/CTA/index.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import './CTA.css'
-
-const CTA = () => {
-  return (
-    <div className='cta_container'>
-        <div>
-            <h3>Always be <br /> inspired</h3>
-            <p>Receive updates about new <br /> and best design posts.</p>
-            <button className='btn'>Join WeDesign</button>
-        </div>
-    </div>
-  )
-}
-
-export default CTA
