+++ src/pages/HomePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import CTA from './../Components/CTA/index'
+import CTA from '../Components/CTA/CTA'
 import Footer from '../Components/Footer/Footer'
 import ImageSlider from './../Components/ImageSlider/ImageSlider'
